moved the logic for matching mutants in the json report and in memory mutants to seperate class to increase testability
