+++ src/Stryker.Core/Stryker.Core.UnitTest/MutantFilters/DashboardMutantFilterTests.cs
@@ -4,9 +4,8 @@ using System.Linq;
 using System.Threading.Tasks;
 using Moq;
 using Shouldly;
-using Stryker.Core.Baseline;
+using Stryker.Core.Baseline.Providers;
 using Stryker.Core.DashboardCompare;
-using Stryker.Core.DiffProviders;
 using Stryker.Core.MutantFilters;
 using Stryker.Core.Mutants;
 using Stryker.Core.Options;
@@ -30,120 +29,6 @@ namespace Stryker.Core.UnitTest.MutantFilters
 
             // Assert
             target.DisplayName.ShouldBe("dashboard filter");
-        }
-
-        [Fact]
-        public void GetMutantSourceShouldReturnMutantSource()
-        {
-            // Arrange
-            var path = $"TestResources{Path.DirectorySeparatorChar}ExampleSourceFile.cs";
-
-            var file = new FileInfo(path);
-
-            var source = File.ReadAllText(file.FullName);
-
-            var gitInfoProvider = new Mock<IGitInfoProvider>(MockBehavior.Loose);
-
-            var baselineProvider = new Mock<IBaselineProvider>(MockBehavior.Loose);
-
-            var jsonMutant = new JsonMutant
-            {
-                Location = new JsonMutantLocation(new JsonMutantPosition
-                {
-                    Column = 17,
-                    Line = 17
-                },
-                new JsonMutantPosition
-                {
-                    Column = 62,
-                    Line = 17
-                }),
-            };
-
-            var target = new DashboardMutantFilter(new StrykerOptions(diff: false), baselineProvider.Object, gitInfoProvider.Object);
-
-            // Act
-            var result = target.GetMutantSourceCode(source, jsonMutant);
-
-            // Assert
-            result.ShouldBe("return Fibonacci(b, a + b, counter + 1, len);");
-        }
-
-        [Fact]
-        public void GetMutantSourceShouldReturnMutantSource_When_Multiple_Lines()
-        {
-            // Arrange
-            var path = $"TestResources{Path.DirectorySeparatorChar}ExampleSourceFile.cs";
-
-            var file = new FileInfo(path);
-
-            var source = File.ReadAllText(file.FullName);
-
-            var gitInfoProvider = new Mock<IGitInfoProvider>(MockBehavior.Loose);
-
-            var baselineProvider = new Mock<IBaselineProvider>(MockBehavior.Loose);
-
-            var jsonMutant = new JsonMutant
-            {
-                Location = new JsonMutantLocation(new JsonMutantPosition
-                {
-                    Column = 13,
-                    Line = 24
-                },
-                new JsonMutantPosition
-                {
-                    Column = 38,
-                    Line = 26
-                }),
-            };
-
-            var target = new DashboardMutantFilter(new StrykerOptions(diff: false), baselineProvider.Object, gitInfoProvider.Object);
-
-            // Act
-            var result = target.GetMutantSourceCode(source, jsonMutant);
-
-            // Assert
-            result.ShouldBe(@"return @""Lorem Ipsum
-                    Dolor Sit Amet
-                    Lorem Dolor Sit"";");
-        }
-
-        [Fact]
-        public void GetMutantSource_Gets_Partial_Line()
-        {
-            // Arrange
-            var path = $"TestResources{Path.DirectorySeparatorChar}ExampleSourceFile.cs";
-
-            var file = new FileInfo(path);
-
-            var source = File.ReadAllText(file.FullName);
-
-            var gitInfoProvider = new Mock<IGitInfoProvider>(MockBehavior.Loose);
-
-            var baselineProvider = new Mock<IBaselineProvider>(MockBehavior.Loose);
-
-            var jsonMutant = new JsonMutant
-            {
-                Location = new JsonMutantLocation(new JsonMutantPosition
-                {
-                    Column = 30,
-                    Line = 34
-                },
-                new JsonMutantPosition
-                {
-                    Column = 34,
-                    Line = 34
-                }),
-            };
-
-            var target = new DashboardMutantFilter(new StrykerOptions(diff: false), baselineProvider.Object, gitInfoProvider.Object);
-
-            // Act
-            var result = target.GetMutantSourceCode(source, jsonMutant);
-
-            // Assert
-            result.ShouldBe("\"\\n\"");
-
         }
 
         [Fact]
