+++ src/Stryker.Core/Stryker.Core/Baseline/Providers/IBaselineProvider.cs
@@ -1,7 +1,7 @@
-﻿using Stryker.Core.Reporters.Json;
+using Stryker.Core.Reporters.Json;
 using System.Threading.Tasks;
 
-namespace Stryker.Core.Baseline
+namespace Stryker.Core.Baseline.Providers
 {
     public interface IBaselineProvider
     {
