+++ src/Stryker.Core/Stryker.Core/MutantFilters/MutantFilterFactory.cs
@@ -1,9 +1,9 @@
-using Stryker.Core.Baseline;
+using System;
+using System.Collections.Generic;
+using Stryker.Core.Baseline.Providers;
 using Stryker.Core.DashboardCompare;
 using Stryker.Core.DiffProviders;
 using Stryker.Core.Options;
-using System;
-using System.Collections.Generic;
 
 namespace Stryker.Core.MutantFilters
 {
